# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorInfo:
    description: Detailed information about an error that occurred during an API request.
    properties:
      code:
        description: |-
          Code is a machine-readable error code identifying the specific error type.
          required: true
          example: RESOURCE_NOT_FOUND
        example: RESOURCE_NOT_FOUND
        type: string
      details:
        description: |-
          Details provides optional additional information about the error, such as validation failures.
          example: {"field": "command", "error": "cannot be empty"}
      message:
        description: |-
          Message is a human-readable description of the error.
          required: true
          example: The requested container was not found.
        example: The requested container was not found.
        type: string
    type: object
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse:
    description: Standard structure for returning errors from the API.
    properties:
      error:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorInfo'
        description: |-
          Error contains the detailed error information.
          required: true
      meta:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
        description: |-
          Meta contains metadata about the response.
          required: true
      success:
        description: |-
          Success indicates if the request was successful (always false for errors).
          required: true
          example: false
        example: false
        type: boolean
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginationResponse'
      request_id:
        example: req-12345
        type: string
      timestamp:
        example: "2023-10-27T10:30:00Z"
        type: string
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkConnectRequest:
    description: Specifies the container and optional endpoint settings for connecting
      to a network.
    properties:
      container:
        description: |-
          Container is the ID or name of the container to connect.
          required: true
          example: my-web-container
        example: my-app-container
        type: string
      endpoint_config:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.EndpointSettingsRequest'
        description: EndpointConfig provides custom network settings for the container
          within this network.
    required:
    - container
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkCreateRequest:
    type: object
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkDisconnectRequest:
    description: Specifies the container to disconnect from a network.
    properties:
      container:
        description: |-
          Container is the ID or name of the container to disconnect.
          required: true
          example: my-web-container
        example: my-app-container
        type: string
      force:
        description: |-
          Force disconnects the container even if it is running.
          example: false
        example: false
        type: boolean
    required:
    - container
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginatedResponse:
    properties:
      data:
        description: Should hold the slice of items
      meta:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse:
    properties:
      data:
        description: Use omitempty if data might be nil/empty
      message:
        description: Optional success message
        type: string
      meta:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Docker Server Manager Go MCP.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Docker Server Manager Go MCP API
  version: "1.0"
paths:
...
  /networks:
    get:
      description: Retrieves a list of Docker networks.
      parameters:
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        example: 20
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - default: created_at
        description: Field to sort by (e.g., name, created_at, driver)
        example: name
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order (asc, desc)
        enum:
        - asc
        - desc
        example: asc
        in: query
        name: sort_order
        type: string
      - description: Filter by driver name
        example: bridge
        in: query
        name: driver
        type: string
      - description: Filter by network name
        example: app-net
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved networks
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkResponse'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List networks
      tags:
      - Networks
    post:
      consumes:
      - application/json
      description: Creates a new Docker network.
      parameters:
      - description: Network Configuration
        in: body
        name: network
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created network
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Network name already exists
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new network
      tags:
      - Networks
  /networks/{id}:
    delete:
      description: Removes a Docker network by its ID or name.
      parameters:
      - description: Network ID or Name
        example: my-app-network
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Network removed successfully
        "400":
          description: Invalid network ID/name
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Network not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Conflict (e.g., network is in use)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a network
      tags:
      - Networks
    get:
      description: Retrieves detailed information about a specific network by its
        ID or name.
      parameters:
      - description: Network ID or Name
        example: my-app-network
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved network details
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkResponse'
              type: object
        "400":
          description: Invalid network ID/name
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Network not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get network details
      tags:
      - Networks
  /networks/{id}/connect:
    post:
      consumes:
      - application/json
      description: Connects a container to a specified network.
      parameters:
      - description: Network ID or Name
        example: my-app-network
        in: path
        name: id
        required: true
        type: string
      - description: Container Connect Info
        in: body
        name: connect_info
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Container connected successfully
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Network or container not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Container already connected
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Connect a container to a network
      tags:
      - Networks
  /networks/{id}/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnects a container from a specified network.
      parameters:
      - description: Network ID or Name
        example: my-app-network
        in: path
        name: id
        required: true
        type: string
      - description: Container Disconnect Info
        in: body
        name: disconnect_info
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.NetworkDisconnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Container disconnected successfully
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Network or container not found, or container not connected
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disconnect a container from a network
      tags:
      - Networks
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"